# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server for create, read and delete tasks",
        "title": "Test Task I/O bound",
        "contact": {},
        "version": "0.6.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/task/delete": {
            "delete": {
                "description": "Delete finished task by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Delete task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponceId"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/restutils.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/restutils.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/task/info": {
            "get": {
                "description": "Get task status, date of creation and processing time by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Get task info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/taskManager.TaskInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/restutils.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/task/new": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Create a new task",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponceId"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ResponceId": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "restutils.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "taskManager.TaskInfo": {
            "type": "object",
            "properties": {
                "create_date": {
                    "type": "string"
                },
                "process_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}